+++ src/components/profile-orders/profile-orders.js
@@ -1,0 +1,9 @@
+import styles from "./profile-orders.module.css";
+
+export const ProfileOrders = () => {
+  return (
+    <div className={`${styles.feed} custom-scroll`}>
+
+    </div>
+  )
+}
