+++ src/components/profile-orders/profile-orders.module.css
@@ -1,0 +1,11 @@
+.feed {
+  border: red solid 1px;
+  width: 850px;
+  height: 700px;
+  margin-right: -20vw;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  align-items: center;
+  overflow-y: scroll;
+}
